chore: Rename ProposalScreenAPI to ProposalDetailScreenAPI
